+++ packages/video_player_avplay/lib/video_player.dart
@@ -15,7 +15,6 @@ import 'package:flutter_tizen/flutter_tizen.dart';
 import 'src/closed_caption_file.dart';
 import 'src/drm_configs.dart';
 import 'src/hole.dart';
-import 'src/subtitle_attribute.dart';
 import 'src/tracks.dart';
 import 'video_player_platform_interface.dart';
 
@@ -64,7 +63,6 @@ class VideoPlayerValue {
     this.playbackSpeed = 1.0,
     this.errorDescription,
     this.isCompleted = false,
-    this.subtitleAttributes = const <SubtitleAttribute>[],
   });
 
   /// Returns an instance for a video that hasn't been loaded.
@@ -143,9 +141,6 @@ class VideoPlayerValue {
   /// Indicates whether or not the video has been loaded and is ready to play.
   final bool isInitialized;
 
-  /// The attributes of the subtitle.
-  final List<SubtitleAttribute> subtitleAttributes;
-
   /// Indicates whether or not the video is in an error state. If this is true
   /// [errorDescription] should have information about the problem.
   bool get hasError => errorDescription != null;
@@ -185,7 +180,6 @@ class VideoPlayerValue {
     double? playbackSpeed,
     String? errorDescription = _defaultErrorDescription,
     bool? isCompleted,
-    List<SubtitleAttribute>? subtitleAttributes,
   }) {
     return VideoPlayerValue(
       duration: duration ?? this.duration,
@@ -206,7 +200,6 @@ class VideoPlayerValue {
               ? errorDescription
               : this.errorDescription,
       isCompleted: isCompleted ?? this.isCompleted,
-      subtitleAttributes: subtitleAttributes ?? this.subtitleAttributes,
     );
   }
 
@@ -227,8 +220,7 @@ class VideoPlayerValue {
         'volume: $volume, '
         'playbackSpeed: $playbackSpeed, '
         'errorDescription: $errorDescription, '
-        'isCompleted: $isCompleted), '
-        'subtitleAttributes: $subtitleAttributes,';
+        'isCompleted: $isCompleted),';
   }
 
   @override
@@ -250,8 +242,7 @@ class VideoPlayerValue {
           volume == other.volume &&
           playbackSpeed == other.playbackSpeed &&
           errorDescription == other.errorDescription &&
-          isCompleted == other.isCompleted &&
-          subtitleAttributes == other.subtitleAttributes;
+          isCompleted == other.isCompleted;
 
   @override
   int get hashCode => Object.hash(
@@ -270,7 +261,6 @@ class VideoPlayerValue {
     playbackSpeed,
     errorDescription,
     isCompleted,
-    subtitleAttributes,
   );
 }
 
@@ -466,58 +456,6 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
     }
   }
 
-  /// Determine the [SubtitleAttrValueType] based on the value of attrTypeNum.
-  static SubtitleAttrValueType _getAttrValueType(int attrTypeNum) {
-    final Set<int> doubleValues = <int>{
-      0,
-      1,
-      2,
-      3,
-      4,
-      5,
-      11,
-      14,
-      19,
-      20,
-      27,
-      31,
-      32,
-    };
-    final Set<int> intValues = <int>{
-      6,
-      7,
-      8,
-      9,
-      10,
-      12,
-      15,
-      16,
-      17,
-      18,
-      21,
-      22,
-      23,
-      24,
-      25,
-      28,
-      29,
-      30,
-      33,
-      34,
-      35,
-    };
-    final Set<int> stringValues = <int>{13, 26};
-    if (doubleValues.contains(attrTypeNum)) {
-      return SubtitleAttrValueType.double;
-    } else if (intValues.contains(attrTypeNum)) {
-      return SubtitleAttrValueType.int;
-    } else if (stringValues.contains(attrTypeNum)) {
-      return SubtitleAttrValueType.String;
-    } else {
-      return SubtitleAttrValueType.none;
-    }
-  }
-
   /// Attempts to open the given [dataSource] and load metadata about the video.
   Future<void> initialize() async {
     await _checkPlatformAndApiVersion();
@@ -628,30 +566,22 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
         case VideoEventType.bufferingEnd:
           value = value.copyWith(isBuffering: false);
         case VideoEventType.subtitleUpdate:
-          final Caption caption = Caption(
-            number: 0,
-            start: value.position,
-            end: value.position + (event.duration?.end ?? Duration.zero),
-            text: event.text ?? '',
-          );
-          value = value.copyWith(caption: caption);
-        case VideoEventType.subtitleAttrUpdate:
           final List<SubtitleAttribute> subtitleAttributes =
               <SubtitleAttribute>[];
           final List<Map<Object?, Object?>?> subtitleAttrList =
-              event.attributes!.cast<Map<Object?, Object?>?>();
+              event.subtitleAttributes!.cast<Map<Object?, Object?>?>();
           for (final Map<Object?, Object?>? attr in subtitleAttrList) {
             final int attrTypeNum = attr!['attrType']! as int;
             final int startTime = attr['startTime']! as int;
             final int stopTime = attr['stopTime']! as int;
             Object attrValue;
-            if (_getAttrValueType(attrTypeNum) ==
+            if (SubtitleAttrType.getValueType(attrTypeNum) ==
                 SubtitleAttrValueType.double) {
               attrValue = attr['attrValue']! as double;
-            } else if (_getAttrValueType(attrTypeNum) ==
+            } else if (SubtitleAttrType.getValueType(attrTypeNum) ==
                 SubtitleAttrValueType.int) {
               attrValue = attr['attrValue']! as int;
-            } else if (_getAttrValueType(attrTypeNum) ==
+            } else if (SubtitleAttrType.getValueType(attrTypeNum) ==
                 SubtitleAttrValueType.String) {
               attrValue = attr['attrValue']! as String;
             } else {
@@ -666,7 +596,14 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
               ),
             );
           }
-          value = value.copyWith(subtitleAttributes: subtitleAttributes);
+          final Caption caption = Caption(
+            number: 0,
+            start: value.position,
+            end: value.position + (event.duration?.end ?? Duration.zero),
+            text: event.text ?? '',
+            subtitleAttributes: subtitleAttributes,
+          );
+          value = value.copyWith(caption: caption);
         case VideoEventType.isPlayingStateUpdate:
           if (event.isPlaying ?? false) {
             value = value.copyWith(