+++ packages/video_player_avplay/lib/video_player_platform_interface.dart
@@ -473,7 +473,7 @@ class VideoEvent {
     this.buffered,
     this.text,
     this.isPlaying,
-    this.attributes,
+    this.subtitleAttributes,
   });
 
   /// The type of the event.
@@ -505,7 +505,7 @@ class VideoEvent {
   final bool? isPlaying;
 
   /// Attributes of the video subtitle.
-  final List<dynamic>? attributes;
+  final List<dynamic>? subtitleAttributes;
 
   @override
   bool operator ==(Object other) {
@@ -550,9 +550,6 @@ enum VideoEventType {
 
   /// Updated the video subtitle text.
   subtitleUpdate,
-
-  /// Updated the video subtitle attributes.
-  subtitleAttrUpdate,
 
   /// The playback state of the video has changed.
   ///